gestion panier

+++ database/migrations/2024_09_12_220814_create_users_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string('name');
             $table->string('profile');
             $table->string('adresse');
-            $table->bigInteger('telephone');
+            $table->integer('telephone');
             $table->foreignIdFor(secteur::class)->constrained()->onDelete('cascade');
             $table->softDeletes('deleted_at', precision: 0);
             $table->string('email')->unique();
